refactor: move get_repositories_safely to get_repos_gql module
Move get_repositories_safely function from main.py to get_repos_gql.py to group related repository functionality together. This improves code organization by keeping all repository retrieval logic in one module.
